Add support to Azure OpenAI
